chore: 取消npz/wav大文件追踪并加入gitignore，提交cogevo条件B跨模型实验与论文初稿
- .gitignore 新增 *.npz *.wav *.pt *.pth

- 移除索引中全部 .npz/.wav

- 新增条件B反馈极性对照、E9/E13-E15实验报告与数据

- 新增论文大纲与初稿（增量记忆装置与外部刺激驱动的行为调整）

- personality.py: 性格段提示改为随使用而保持稳定

- schemanet: 提交E6/E7补充实验、论文草稿与答辩复述清单

+++ schemanet/snapshot.py
@@ -107,7 +107,7 @@ def _restore_net(z):
         src_i, slot_k, dst_j, vals = z["src_i"], z["slot_k"], z["dst_j"], z["vals"]
         _rows_from_arrays(ng, src_i, slot_k, dst_j, vals)   # 批量构建（免逐条 dict 插入）
         if "gain" in z:   # 旧快照无 gain 字段 → 保持默认全 1（向后兼容）
-            ng.gain = z["gain"].astype(np.float64)
+            ng.gain = z["gain"].astype(np.float32)   # 统一 float32 域（860 亿规模内存优化）
     if "skeletons" in z and z["skeletons"].tobytes():
         # json 序列化把 int 键变字符串——恢复时转回 int（fixed/content/bound）
         def _fix_keys(d):
